added intake air linear service

+++ HardwareAbstraction/HardwareAbstraction/IAnalogService.h
@@ -4,6 +4,6 @@ namespace HardwareAbstraction
 	{
 	public:
 		virtual void InitPin(uint8_t pin) = 0;
-		virtual unsigned int ReadPin(uint8_t pin) = 0;
+		virtual unsigned short ReadPin(uint8_t pin) = 0;
 	};
 }
